requirements.txt
mkdocs-material==9.4.7
Django~=4.2.7
djangorestframework~=3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-extensions==3.2.3
djoser==2.2.2
drf-yasg==1.21.7